refactor(security): 安全策略面板拆成危险模式 / SSRF 两个 Tab
SecurityPanel 瘦身成 Tabs wrapper：
- Tab 1 "危险模式" → DangerousModeSection（即时 toggle，无底部保存按钮）
- Tab 2 "SSRF 策略" → 新抽出的 SsrfPolicySection（保留原有 default policy / per-tool / trusted hosts 三块 + sticky 保存按钮）

DangerousModeSection 内重复的"危险模式"标题移除（Tab 标题已自述）。
新增 i18n settings.tabDangerous / settings.tabSsrf。
