+++ src/i18n/locales/en.json
@@ -363,6 +363,8 @@
     "tabAppearance": "Appearance & Language",
     "tabSystem": "System & Shortcuts",
     "tabNetwork": "Network Proxy",
+    "tabDangerous": "Dangerous Mode",
+    "tabSsrf": "SSRF Policy",
     "dangerousTitle": "Dangerous Mode",
     "dangerousIntro": "One-click switch to bypass every tool approval. Extreme risk — enable only on fully trusted local machines.",
     "dangerousSwitchLabel": "Skip all tool approvals",
